combat/requirements.txt
crawl4ai>=0.2.0
scrapy>=2.11.0
markdownify>=0.11.0
httpx>=0.25.0
tabulate>=0.9.0
pytest-asyncio>=0.23.0
